Converting CalChart to use CMake.
This makes it easier to work on both Mac and Windows.
Adding wxWidgets as a submodule and part of the build system to remove an extra step.
Integrating with cpack to make an installer step.
