--- build-win/CalChart/CalChart.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Express 2013 for Windows Desktop
-VisualStudioVersion = 12.0.21005.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CalChart", "CalChart.vcxproj", "{17397087-23C1-424D-9796-EFD951CDFBB0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Release|Win32 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{17397087-23C1-424D-9796-EFD951CDFBB0}.Debug|Win32.ActiveCfg = Debug|Win32
-		{17397087-23C1-424D-9796-EFD951CDFBB0}.Debug|Win32.Build.0 = Debug|Win32
-		{17397087-23C1-424D-9796-EFD951CDFBB0}.Release|Win32.ActiveCfg = Release|Win32
-		{17397087-23C1-424D-9796-EFD951CDFBB0}.Release|Win32.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
