+++ src/animation_frame.cpp
@@ -656,3 +656,4 @@ void AnimationFrame::OnNotifyErrorList(
         mErrorMarkers.emplace_back(i.second, sheetnum);
     }
 }
+
